fix: restaura XMLAbstracts em sps/models, corrige categorizacao da PR #1180
Atende a revisao na PR #1334 (issue #1329): apontou que a correcao de
idioma do resumo de sub-article foi feita em
sps/validation/models/abstract.py, mas essa classe nao e exclusiva de
validacao.

Confirmado: XMLAbstracts foi movida de sps/models/v2/abstract.py para
sps/validation/models/abstract.py na PR #1180 (maio/2026, Copilot) sob a
premissa de que era usada so por validacao. Essa premissa estava errada:
scms-upload importa XMLAbstracts via packtools.sps.models.v2.abstract
(document.py) para montar dados do TOC do fasciculo, nao para validar
nada. So funcionava porque o modulo antigo virou um shim de
compatibilidade redirecionando para a classe em validation/models.

A classe em si nunca teve logica de validacao (so extracao/organizacao
de dados via xpath, sem comparacao contra regra ou geracao de
advice/erro), entao nunca deveria ter sido categorizada como exclusiva
de validacao.

Mudancas:
- packtools/sps/models/v2/abstract.py: implementacao real de
  XMLAbstracts restaurada (com o fix do #1329 ja aplicado: idioma do
  abstract de sub-article nunca cai para o lang do artigo principal).
  Remove o shim de deprecacao que redirecionava para validation/models.
- packtools/sps/validation/models/abstract.py: vira o shim, na direcao
  oposta (redireciona para models/v2, nao o contrario), documentado como
  compatibilidade apenas, sem novo consumidor conhecido (nenhum dos 10
  repos scieloorg dependentes de packtools importa desse caminho).
- packtools/sps/validation/article_abstract.py: import atualizado para
  o caminho correto (models/v2), evitando warning de deprecacao interno.

Fecha as issues #1153 e #1181 (separacao de responsabilidades entre
modelos de dados e validacao: abstract). Nao mexe nas outras 14 classes
que a PR #1180 tambem moveu, fica como recomendacao separada.

Suite completa: 6139 passed, 31 skipped, 461 subtests passed.
